+++ 16.12.01 Portfolio Framework/Dx3D/cUIImageView.cpp
@@ -15,7 +15,7 @@ cUIImageView::~cUIImageView(void)
 void cUIImageView::Render(LPD3DXSPRITE pSprite)
 {
 	RECT rc;
-	SetRect(&rc, 0, 0, m_stSize.fWidth, m_stSize.fHeight);
+	SetRect(&rc, 0, 0, (int)m_stSize.fWidth, (int)m_stSize.fHeight);
 	pSprite->Begin(D3DXSPRITE_ALPHABLEND | D3DXSPRITE_SORT_TEXTURE);
 	pSprite->SetTransform(&m_matWorld);
 	pSprite->Draw(m_pTexture,
